+++ packages/components/src/Description/DescriptionContext/DescriptionContextProvider/DescriptionContextProvider.tsx
@@ -1,0 +1,20 @@
+import { type ReactNode } from 'react';
+
+import {
+  DescriptionContext,
+  type DescriptionContextProps,
+} from '../DescriptionContext';
+
+export type DescriptionProviderProps = DescriptionContextProps & {
+  children: ReactNode;
+};
+
+export const DescriptionContextProvider = ({
+  children,
+  leader,
+  separator,
+}: DescriptionProviderProps) => (
+  <DescriptionContext.Provider value={{ leader, separator }}>
+    {children}
+  </DescriptionContext.Provider>
+);
