+++ packages/components/src/Description/DescriptionContext/DescriptionContextProvider/index.ts
@@ -1,0 +1,1 @@
+export * from './DescriptionContextProvider';
