+++ packages/components/src/Description/DescriptionContext/index.ts
@@ -1,0 +1,3 @@
+export * from './DescriptionContext';
+
+export * from './DescriptionContextProvider';
