--- packages/components/src/Description/DescriptionProvider/index.ts
@@ -1,1 +1,0 @@
-export * from './DescriptionProvider';
